% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALDEx2jlr.R
\name{aldex2gpu_setup}
\alias{aldex2gpu_setup}
\title{Setup ALDEx2GPU.jl}
\usage{
aldex2gpu_setup(pkg_check = TRUE, ...)
}
\arguments{
\item{pkg_check}{logical, check for ALDEx2GPU.jl package and install if necessary}

\item{...}{Parameters are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the ALDEx2GPU.jl package.
The first time will be long since it includes precompilation.
Additionally, this will install Julia and the required packages
if they are missing.
}
\examples{
\dontrun{
## aldex2gpu_setup() is time-consuming and requires Julia + ALDEx2GPU.jl

aldex2jlr::aldex2gpu_setup()
}

}
