% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALDEx2jlr.R
\name{aldex2jl_setup}
\alias{aldex2jl_setup}
\title{Setup ALDEx2jlr}
\usage{
aldex2jl_setup(pkg_check = TRUE, ...)
}
\arguments{
\item{pkg_check}{logical, check for ALDEx2.jl package and install if necessary}

\item{...}{Parameters are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the ALDEx2.jl package.
The first time will be long since it includes precompilation.
Additionally, this will install Julia and the required packages
if they are missing.
}
\examples{
\dontrun{
## aldex2jlr_setup() is time-consuming and requires Julia + ALDEx2.jl

aldex2jlr::aldex2jl_setup()
}

}
